jbdc+postgresql+liquibase: big/small game now saves and reads from db :D

+++ src/main/java/ee/eek/bigsmall/dto/NewGameResponse.java
@@ -6,5 +6,5 @@ import lombok.experimental.Accessors;
 @Data
 @Accessors(chain = true)
 public class NewGameResponse {
-    private Integer gameID;
+    private Long gameID;
 }
